+++ CotEditor/Sources/WindowSettingsView.swift
@@ -27,6 +27,8 @@ import SwiftUI
 
 struct WindowSettingsView: View {
     
+    @Environment(\.layoutDirection) private var layoutDirection
+    
     @AppStorage(.windowTabbing) private var windowTabbing
     @AppStorage(.windowWidth) private var windowWidth
     @AppStorage(.windowHeight) private var windowHeight
@@ -199,7 +201,7 @@ struct WindowSettingsView: View {
                 
                 Stepper(value: $overscrollRate, in: 0...1, step: 0.1, format: .percent.precision(.fractionLength(0)), label: EmptyView.init)
                     .monospacedDigit()
-                    .environment(\.layoutDirection, .rightToLeft)
+                    .multilineTextAlignment(self.layoutDirection == .rightToLeft ? .leading : .trailing)
             }
             
             Divider()