+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixes
 
 - Fix an issue on CotEditor 4.8.0 that the editor's kerning did not set to zero despite the font type is monospaced.
+- Fix an issue on CotEditor 4.8.0 that some steppers in the Settings window were placed on the opposite side.